Fix blank screen issue in workspace view
Fixed runtime error in ConfigMatrix component where "Show All" and "Hide All"
dropdown menu items were using DropdownMenuCheckboxItem with onCheckedChange,
but the handler functions don't accept parameters. Changed to DropdownMenuItem
with onClick which is the correct pattern for action buttons.

This was causing the workspace view to fail rendering after selecting a project.